Move code to SystemModuleManager
